Add beneficiary to campaign details page - inital changes
